better debug messages
When things don't work it helps to have commands
in the log how to reproduce the BenchmarkAnything
queries manually with curl.

Also, as quite extreme measures, we print out the
results in debug mode. That might be a bit too
much, though. In that case we drop it again later.